refactor: move ParallaxText to the /common folder
