+++ src/components/Governance/ParallaxDaoStats/ParallaxDaoStats.tsx
@@ -1,5 +1,5 @@
 import ParallaxDaoStatsElement from '@/components/Governance/ParallaxDaoStats/ParallaxDaoStatsElement'
-import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxTextOld'
 
 const ParallaxDaoStats = (props: ParallaxTextProps) => {
   return (
