+++ src/components/Wallet/ParallaxBatching/index.tsx
@@ -1,5 +1,5 @@
 import ParallaxBatchingElement from '@/components/Wallet/ParallaxBatching/ParallaxBatchingElement'
-import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxTextOld'
 
 const ParallaxBatching = (props: ParallaxTextProps) => {
   return (
