+++ src/components/Wallet/ParallaxTxSimulation/index.tsx
@@ -1,5 +1,5 @@
 import ParallaxTxSimulationElement from '@/components/Wallet/ParallaxTxSimulation/ParallaxTxSimulationElement'
-import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxText'
+import ParallaxText, { type ParallaxTextProps } from '@/components/common/ParallaxTextOld'
 
 const ParallaxTxSimulation = (props: ParallaxTextProps) => {
   return (
