+++ src/components/common/ParallaxText/index.tsx
@@ -1,50 +1,47 @@
 import type { ReactNode } from 'react'
 import { Container, Grid, Typography } from '@mui/material'
+import clsx from 'clsx'
+import { type BaseBlockEntry } from '@/config/types'
+import RichText from '@/components/common/RichText'
+import ButtonsWrapper from '@/components/Token/ButtonsWrapper'
+import { isEntryTypeButton } from '@/lib/typeGuards'
 import layoutCss from '@/components/common/styles.module.css'
 import css from './styles.module.css'
-import type { BaseBlock } from '@/components/Home/types'
-import Stepper, { type StepsType } from '@/components/Wallet/Stepper'
-import clsx from 'clsx'
-import ButtonsWrapper from '@/components/common/ButtonsWrapper'
 
-export type ParallaxTextProps = BaseBlock & {
+export type ParallaxTextProps = BaseBlockEntry & {
   variant: 'image-text' | 'text-image'
-  mobileVariant?: 'image-text' | 'text-image'
+  children: ReactNode
 }
 
-const ParallaxText = ({
-  title,
-  text,
-  buttons,
-  steps,
-  caption,
-  variant,
-  mobileVariant = 'image-text',
-  children,
-}: ParallaxTextProps & { children: ReactNode; steps?: StepsType }) => {
+const ParallaxText = (props: ParallaxTextProps) => {
+  const { caption, title, text, buttons } = props.fields
+  const { variant, children } = props
+
+  const buttonsList = buttons?.filter(isEntryTypeButton) || []
+
   return (
     <Container>
       <Grid
         container
-        className={clsx(
-          layoutCss.containerMedium,
-          variant === 'image-text' ? css.imageFirst : css.textFirst,
-          mobileVariant === 'text-image' ? css.textFirstMobile : null,
-        )}
+        className={clsx(layoutCss.containerMedium, variant === 'image-text' ? css.imageFirst : css.textFirst)}
         spacing={{ xs: 6, md: '30px', xl: '50px' }}
         justifyContent="space-between"
       >
-        <Grid item xs={12} md={5} display="flex" flexDirection="column" justifyContent="center" gap={{ xs: 3, md: 4 }}>
+        <Grid item xs={12} md={6} className={css.textBlockWrapper}>
           {caption && (
-            <Typography variant="caption" component="div">
+            <Typography variant="caption" component="div" mb="32px">
               {caption}
             </Typography>
           )}
-          <Typography variant="h2">{title}</Typography>
-          <Typography className={css.textBlock}>{text}</Typography>
-          {steps && <Stepper steps={steps} />}
-          <ButtonsWrapper buttons={buttons} />
+
+          <div>
+            <RichText {...title} />
+
+            {text && <RichText {...text} />}
+          </div>
+          <ButtonsWrapper buttons={buttonsList} />
         </Grid>
+
         <Grid item xs={12} md={6} display="flex" alignItems="center">
           {children}
         </Grid>
